refactor: Move to adapter.
