+++ filejacket/adapters/storage.py
@@ -40,7 +40,7 @@ from pathlib import (
 # third-party
 from typing import Any, Pattern
 
-from .engines.storage import StorageEngine
+from ..engines.storage import StorageEngine
 
 
 __all__ = [
