+++ filejacket/image/__init__.py
@@ -20,15 +20,3 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 Should there be a need for contact the electronic mail
 `filejacket <at> gabrielfontenelle.com` can be used.
 """
-
-from .engine import ImageEngine
-from .engine.opencv import OpenCVImage
-from .engine.pillow import PillowImage
-from .engine.wand import WandImage
-
-__all__ = [
-    "ImageEngine",
-    "OpenCVImage",
-    "PillowImage",
-    "WandImage"
-]
